Move JsonConfigurationJob to Jobs.Common
