+++ src/ArchivePackages/ArchivePackages.csproj
@@ -44,6 +44,10 @@
   <ItemGroup>
     <Compile Include="ArchivePackages.Job.cs" />
     <Compile Include="ArchivePackages.Program.cs" />
+    <Compile Include="Configuration\InitializationConfiguration.cs" />
+    <Compile Include="JobEventSource.cs" />
+    <Compile Include="JobTasks.cs" />
+    <Compile Include="PackageRef.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -52,6 +56,9 @@
     </None>
     <None Include="Scripts\*" />
     <None Include="ArchivePackages.nuspec" />
+    <None Include="Settings\dev.json" />
+    <None Include="Settings\int.json" />
+    <None Include="Settings\prod.json" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\NuGet.Jobs.Common\NuGet.Jobs.Common.csproj">
@@ -75,9 +82,6 @@
     <PackageReference Include="Newtonsoft.Json">
       <Version>9.0.1</Version>
     </PackageReference>
-    <PackageReference Include="NuGet.Services.Sql">
-      <Version>2.27.0</Version>
-    </PackageReference>
     <PackageReference Include="System.Net.Http">
       <Version>4.3.3</Version>
     </PackageReference>
@@ -85,6 +89,7 @@
       <Version>7.1.2</Version>
     </PackageReference>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="..\..\build\sign.targets" Condition="Exists('..\..\build\sign.targets')" />
 </Project>
