+++ src/Validation.Common.Job/Validation.Common.Job.csproj
@@ -61,7 +61,7 @@
     <Compile Include="Storage\ValidatorStateService.cs" />
     <Compile Include="Storage\ValidatorStatusExtensions.cs" />
     <Compile Include="SubcriptionProcessorJob.cs" />
-    <Compile Include="JsonConfigurationJob.cs" />
+    <Compile Include="ValidationJobBase.cs" />
     <Compile Include="SubscriptionProcessorConfiguration.cs" />
     <Compile Include="Validation\IValidatorProvider.cs" />
     <Compile Include="Validation\ValidatorName.cs" />
